Begin bus passes (#5)
* FEAT: Endpoint to Country Codes

- Added router & controller of country code
- Added controller manager
- Able to:
  - Create
  - List
  - Get
  - Remove
  - Update

* FEAT: Docker compose API-Database

- Added Docker compose
- Added PostgreSQL in docker
- Added datasets
- Updated Project Entry Page
- Reorganized directories

* FEAT: Country Code Controller & Model

- Routes have been improved
- Controller has async locks
- Country Code Controller has all the methods
- Country Code Model finished
- Country Code Controller uses the database instead of memory
- Added a logger
- Added pageable to list
- Fixed country code dataset

* FEAT: Request Validation

- Getting back packet rejection messages
- Added validations to pageables
- Added API docs
- Update Validation Packet

* FEAT: Bus Passes

- Added dataset of bus pass
- Added Model
- Added Controller
- Added Router

* FEAT: Multiple order arguments

- Added possibility of ordering with multiple arguments
- Changed order format

+++ project/business/src/packet_handler/templates/PacketTemplates.cs
@@ -101,8 +101,8 @@ namespace PacketTemplates {
                 bool is_list = (item.Attribute("list")?.Value ?? "FALSE").ToUpper() == "TRUE";
 
                 Type datatype = (item.Attribute("datatype")?.Value ?? "string").ToLower() switch {
-                    "integer" => typeof(int),
-                    "float" => typeof(float),
+                    "integer" => typeof(long),
+                    "float" => typeof(double),
                     "bool" => typeof(bool),
                     _ => typeof(string)
                 };
